+++ src/main/kotlin/me/schlaubi/intellij_gradle_version_checker/inspection/dependencies/inconstant_format/InconstantFormatInspection.kt
@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2020 Michael Rittmeister
+ * Copyright (c) 2020-2022 Michael Rittmeister
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -27,8 +27,6 @@ package me.schlaubi.intellij_gradle_version_checker.inspection.dependencies.inco
 import com.intellij.codeInspection.LocalInspectionToolSession
 import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.psi.PsiElementVisitor
-import io.ktor.client.request.forms.*
-import io.ktor.client.utils.*
 import me.schlaubi.intellij_gradle_version_checker.GradleUpdaterBundle
 import me.schlaubi.intellij_gradle_version_checker.dependency_format.dependencyFormat
 import me.schlaubi.intellij_gradle_version_checker.inspection.AbstractBuildScriptInspection