+++ CHANGELOG.md
@@ -1,17 +1,32 @@
 # IntelliJ Gradle Updater
 
 ## [Unreleased]
+
 ### Changed
+
 - Update to IntelliJ 2021.3
+
+### Fixed
+- Gradle update notification appearing on latest Gradle version
+- Kotlin format converter triggering when converting Gradle snippets
+
 ## [2.0.4]
+
 ### Changed
+
 - Update to IntelliJ 2021.2
+
 ## [2.0.3]
+
 - Fix compatibility range to include all 2021.1 versions
 - Fix compatibility problems
+
 ## [2.0.2]
+
 ### Added
+
 - Add CI
+
 ### Changed
 - Bump dependencies
 ### Fixed